TFL_Streaming/requirements.txt
pyspark==2.4.7
kafka-python==2.1.0
requests==2.27.1
pytest==6.2.5
PyYAML==5.4.1